updates to support ct data uploads
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@5145 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ connect/org.wcs.smart.connect.server/src/main/resources/org/wcs/smart/connect/api/DataQueue.java
@@ -288,14 +288,15 @@ public class DataQueue {
 			ServerDataQueueItem item = (ServerDataQueueItem)s.get(ServerDataQueueItem.class, itemUuid);
 			
 			validateDelete(item.getConservationArea(), s);
-			
-			File toDelete = DataStoreManager.INSTANCE.getFile(item.getFile());
-			if (toDelete.exists()){
-				if (!toDelete.delete()){
-					logger.log(Level.WARNING, "Could not delete data queue file: " + toDelete.toString()); //$NON-NLS-1$
+			if (item.getFile() != null){
+				File toDelete = DataStoreManager.INSTANCE.getFile(item.getFile());
+				if (toDelete.exists()){
+					if (!toDelete.delete()){
+						logger.log(Level.WARNING, "Could not delete data queue file: " + toDelete.toString()); //$NON-NLS-1$
+					}
+				}else{
+					logger.log(Level.WARNING, "Data queue file does not exist to delete: " + toDelete.toString()); //$NON-NLS-1$
 				}
-			}else{
-				logger.log(Level.WARNING, "Data eue file does not exist to delete: " + toDelete.toString()); //$NON-NLS-1$
 			}
 			s.delete(item);
